Lille1

+++ scripts/daily_batch.sh
@@ -1,12 +1,25 @@
 #! /bin/bash
 
+if [ -z "$1" ]
+then
+    DATE=`date -d yesterday +"%Y/%m/%d"`
+else
+    DATE=$1
+fi
+
+
 ES_DIR=/home/elasticsearch
 NBDAY_KEEP_LOG=3
 LOG_DIR=/data/logs
-REP_LOGS=$LOG_DIR/$(date '+%Y%m%d')
+REP_LOGS=$LOG_DIR/$DATE
 ERROR_LOG=$REP_LOGS/error.agimus.log
 INFO_LOG=$REP_LOGS/info.agimus.log
 STAT_LDAP_LOG=$REP_LOGS/stats-ldap.log
+BUILD_HOME="/opt/agimus-ng/build"
+LOGSTASH_DIR="/opt/logstash"
+
+DELETE_OLD_LOG=true
+DUMP_KIBANA_ES=false
 
 MAIL_DEST="agimus-tech@univ.fr"
 
@@ -21,16 +34,20 @@ echo "#### Start of the process : "`date +'%F %R'`
 echo "$LINE_SEPARATOR"
 echo ""
 
-echo "$LINE_SEPARATOR"
-echo "#### Delete old log - $NBDAY_KEEP_LOG days ago : "`date +'%F %R'`
-REP_LOGS_OLD=$LOG_DIR/$(date --date $NBDAY_KEEP_LOG' days ago' '+%Y%m%d')
-rm -f $REP_LOGS_OLD/*.log.{gz,bz2}
-echo ""
+if [ "$DELETE_OLD_LOG" = true ] ; then
+    echo "$LINE_SEPARATOR"
+    echo "#### Delete old log - $NBDAY_KEEP_LOG days ago : "`date +'%F %R'`
+    REP_LOGS_OLD=$LOG_DIR/$(date --date $NBDAY_KEEP_LOG' days ago' '+%Y%m%d')
+    rm -f $REP_LOGS_OLD/*.log.{gz,bz2}
+    echo ""
+fi
 
+if [ "$DUMP_KIBANA_ES" = true ] ; then
 echo "$LINE_SEPARATOR"
-echo "#### Dump Elasticsearch templates and .kibana4 index
+echo "#### Dump Elasticsearch templates and .kibana4 index"
 $ES_DIR/bin/dump_kibana_ES_conf.sh
 echo ""
+fi
 
 echo "$LINE_SEPARATOR"
 echo "#### Delete yesterday index ldap : "`date +'%F %R'`
@@ -40,7 +57,7 @@ echo ""
 
 echo "$LINE_SEPARATOR"
 echo "#### Rebuild of ldap index and make stats index : "`date +'%F %R'`
-logstash --quiet -f logstash/logstash-ldap.conf >&2 && python $ES_DIR/scripts/ldap-agg.py > $STAT_LDAP_LOG
+$LOGSTASH_DIR/bin/logstash --quiet -f $BUILD_HOME/logstash/logstash-ldap.conf >&2 && python $BUILD_HOME/scripts/ldap-agg.py > $STAT_LDAP_LOG
 echo "See : $STAT_LDAP_LOG"
 echo ""
 
@@ -48,7 +65,7 @@ echo "$LINE_SEPARATOR"
 echo "#### Import CAS trace : "`date +'%F %R'`
 if [ -f "$REP_LOGS/trace-cas.log.bz2" ]; then
 	echo "#### Number of lines in file "`bzcat $REP_LOGS/trace-cas.log.bz2 | wc -l`
-	bzcat $REP_LOGS/trace-cas.log.bz2 | logstash --quiet -f logstash/logstash-trace.conf >&2
+	bzcat $REP_LOGS/trace-cas.log.bz2 | $LOGSTASH_DIR/bin/logstash --quiet -f $BUILD_HOME/logstash/logstash-trace.conf >&2
 else 
 	echo "ERR : NO file logs CAS-TRACE" >&2
 fi
@@ -58,7 +75,7 @@ echo "$LINE_SEPARATOR"
 echo "#### Import ENT logs : "`date +'%F %R'`
 if [ -f "$REP_LOGS/access-ent.log.gz" ]; then
 	echo "#### Number of lines in file "`zcat $REP_LOGS/access-ent.log.gz | wc -l`
-	zcat $REP_LOGS/access-ent.log.gz | logstash --quiet -f logstash/logstash-esup4.conf >&2
+	zcat $REP_LOGS/access-ent.log.gz | $LOGSTASH_DIR/bin/logstash --quiet -f $BUILD_HOME/logstash/logstash-esup3.conf >&2
 else
         echo "ERR : NO file logs ENT" >&2
 fi
@@ -68,7 +85,7 @@ echo "$LINE_SEPARATOR"
 echo "#### Import ARCHE logs : "`date +'%F %R'`
 if [ -f "$REP_LOGS/arche-access.log.gz" ]; then
 	echo "#### Number of lines in file "`zcat $REP_LOGS/arche-access.log.gz | wc -l`
-	zcat $REP_LOGS/arche-access.log.gz | logstash --quiet -f logstash/logstash-moodle.conf >&2
+	zcat $REP_LOGS/arche-access.log.gz | $LOGSTASH_DIR/bin/logstash --quiet -f $BUILD_HOME/logstash/logstash-moodle.conf >&2
 else 
         echo "ERR: NO file logs ARCHE" >&2
 fi
